- some refinements to CC (formatting and types)
git-svn-id: https://svn.code.sf.net/p/codeblocks/code/trunk@9043 2a5c6006-c6dd-42ca-98ab-0921f2732cef

+++ src/plugins/codecompletion/insertclassmethoddlg.cpp
@@ -109,7 +109,7 @@ wxArrayString InsertClassMethodDlg::GetCode() const
     wxArrayString array;
     const wxCheckListBox* clb = XRCCTRL(*this, "chklstMethods", wxCheckListBox);
 
-    for (unsigned int i = 0; i < clb->GetCount(); ++i)
+    for (size_t i = 0; i < clb->GetCount(); ++i)
     {
         if (clb->IsChecked(i))
         {
